clean up the cli
