+++ cli/utils/utils.go
@@ -1,0 +1,9 @@
+package utils
+
+func ShortenHex(publicKey string) string {
+	return publicKey[0:6] + ".." + publicKey[len(publicKey)-4:]
+}
+
+// maximum number of proofs per txn for each of the following proof types:
+const DEFAULT_BATCH_CREDENTIALS = 60
+const DEFAULT_BATCH_CHECKPOINT = 80
